refactor: アクセス制御を高階コンポーネントを使った実装に変更
システムの柔軟性を上げるため
